+++ examples/checker.rs
@@ -1,13 +1,13 @@
 extern crate pix;
 
-use pix::{Gray8, Raster, RasterBuilder};
+use pix::{Raster, RasterBuilder, SepSGray8};
 use std::fs::File;
 use std::io;
 use std::io::Write;
 
 fn main() -> Result<(), io::Error> {
-    let v = Gray8::from(255);
-    let mut r = RasterBuilder::<Gray8>::new().with_clear(16, 16);
+    let v = SepSGray8::from(255);
+    let mut r = RasterBuilder::<SepSGray8>::new().with_clear(16, 16);
     for y in 0..16 {
         for x in 0..16 {
             if x + y & 1 != 0 {
@@ -18,7 +18,7 @@ fn main() -> Result<(), io::Error> {
     write_pgm(&r, "checker.pgm")
 }
 
-fn write_pgm(raster: &Raster<Gray8>, filename: &str) -> io::Result<()> {
+fn write_pgm(raster: &Raster<SepSGray8>, filename: &str) -> io::Result<()> {
     let fl = File::create(filename)?;
     let mut bw = io::BufWriter::new(fl);
     let w = bw.get_mut();